refactor(mcp): move channel push logic into the mcpserver package
The xagent-specific channel logic (instructions text, sender
interface, translation, forwarding, and PushChannels entry
point) now lives in internal/server/mcpserver alongside the
user-facing tools. internal/command/mcp.go stays as glue: it
wires the server, advertises the capability via
mcpchannel.Experimental(), and starts the push goroutine via
mcpserver.PushChannels.

internal/x/mcpchannel remains generic — only the Claude Code
channel protocol primitives.
